Corrections and Distance
Corrected the algorithm to add up all the muscles, during one iteration loop, instead of the one that was being changed that iteration. Also added a term to the cost function to account for distance moved away from the starting Location. Also corrected surface plots to account for the best robot torque and the starting Robot torque.