Knee test ssolid models
Solid models for the robotics components of the Knee test. Was originally for a Living Machines 2020 paper, so I updated it to 2021, although we may submit it to another journal. Contains Alex Steel's solid models. Contains part of the air fittings for the puppy model, including ones I made for the 20 mm and 40 mm Festo hoses.  We will rename them and add them into our robot following our naming convention.  For example 11_01_001 for 20 mm end cap.