Solid models for static BPA test
Added folder for BPA static test solid models. Parts are for the frame jig designed by David P. Parts are for 10 mm Festo muscles, and force sensor that was used in David's project. We need similar parts for 20 and 40 mm I.D. BPAs. David reported that deflection is a major design consideration, so the 10 mm can be reinforced with carbon fiber but the larger sizes probably need ribs added at the very least.