Vendor solid models added
Added a folder to put Vendor solid models in. Added Straight and elbow push to connect fittings, so we can try to decrease the inflation time for larger BPAs. Added hose clamps for 20 mm and 40 mm BPA. Added a BoM so it is easy to track what each solid model is and where to find it.