Began working on a new muscle class
Began creating a new muscle class to accomodate the changes I want to make in the algorithm. Namely, to make it so the BPA is only in the longest contiguous muscle segement and to also update the calculations to include the new vector math that was arrived at during the muscle calculations.