Reorganization, finishing biPamClass, beginning comparison
Reorganized the file structure to hide all the previous optimization code and begin establishing a file system for the optimization code. Nearly done with the BiPamClass verification, still need to do some hand calcs to make sure everything is correct. Set up a script to compare the human adductor magnus muscles to the BPA adductor magnus.