Adding test script for force calculation
Trying to elegantly put together a muscle force calculation based on Millard, Hoy, and Thelen. The issue is we need to balance the tendon force and the muscle force. We can do this with built in Matlab solver, but I need to figure out how to make the function maleable enough to also bring in parameters from the muscle during the solving process.