Added updated force calculation
Updated the muscle classes to include a better force calculation, which iterates to find a solution to the muscle tendon force and the passive and elastic muscle force. This was done for the monoarticulate and biarticulate muscle classes. Kind of verified the mono case, want to verify the biarticulate case. After that I should update all of the muscle generators to include the optimal fiber length, pennation angle, and tendon slack length