Changed mono pam class to handle and dependent properties
Changing the mono pam class to be a handle class (one that I can change the values of, such as Location). Also began making the muscle length and other values into dependent values, which are recalculated every time they are querried. However, right now the code bounces between calculating the segment lengths, as segment lengths are called a few times in the script. I need to read more on how to use dependent values to avoid this issue