Updated both PAM classes as handles with dependent variables
Updated the mono and bi pam classes so that they are handles with dependent variables (length, contraction, force direction, and torque, etc) so that they are calculated every time they are querried. This will hopefully make the optimization code easier to navigate.

Try to check the values being calculated when running the optimization code. If there is anything sketchy, we can revert back to when the values are not dependent.