Added line to tell how many calculations will be performed
Added a quick line that displays when the abductor magnus script runs that tells how many calculations will be performed. This is so the user understands the length of time in the calculation.

Will now be creating a separate branch to once again work on a comparison aglorithm and optimization.