Playing around with ways of calculating torque difference
Ran through some simple examples to try to figure out how I wanted to quantify torque differences. When human torque is negative, we also want the PAM torque to be negative, so the algorithm needs to be slightly more nuances than just subtracting human torque from pam torque. I think I settled on a solution that just switches what is subtracted depending on the sign of the human torque. Later on, I will need to decide how to weight and add the torque difference to the cost function.