Finishing mesh calculation and updated monopam data
Did a lot of fighting with the mesh calculation but the bicep femoral muscle code is now working, including both what orientation provides the best torque as well as if the muscle length is usable if it were an actual PAM. I had to also update the usability check in the monopam data, which included editing where the tendon length is calculated. I'll have to update BiPamData as well soon.