Focusing on bifemsh
Focusing on bifemsh so it can be a guide for later muscles. Updated cost function to now have gains and includes a component based on the angle between the vectors. Updated the graphs on the bifemsh script to include the results of PAM that uses the human muscle locations. Created a new graph that plots the angle between vectors.