Creating the bimesh code
Mono articulate muscle code looks good. Working through the biarticulate code. Need to make sure that the transformation matrices are set up correctly for the class to crunch numbers. The class also needs to make sure that it is calculation muscle segment length properly and outputting values in a patterned manner.