Updated Cost Function, via point, refinement code, and multiple add mags
Updated a specific cost function for the knee that penalizes deviation from the z torque more heavily than other axes, since that axis is the one of interest for the knee. When moving that to other joints, I might want an automated way to tell what axis is most important (thinking the adductor muscles want emphasis on the x axis). Also began creating an optimization code for after the mesh portion. Use the mesh solution to search around that area. I did that because for the bifemsh I added a via point and realized that was unnecessary, but what was necessary was getting away from the bone itself, since the OpenSim muscles aren't directly on the bone. Finished creating an Add Mag script that uses 3 bpas instead of 1. Haven't run that script because it will take literal years to finish running.