Created f10 and HuntEq m files
f10 uses a function that was fit to the surface of Dr. Hunt's lookup table. HuntEq solves Dr. Hunt's festo BPA equation for force and can be called by other programs. This was previously in LookupTable_generate but was moved to a separate m-file because there are other times we may want to use it.
2 comments added to MonoPamDataExplicit.
Knee_Flexor_Pinned had human comparison data and plotting removed.
FestoLookup had the save command commented out so that I can do some tests without overriding existing data and fits that seem to work fairly well. Might want to add a 10mm fit based off of the App we got from Festo.