Trying to utilize Matlab's arduino support package
Temperature Logging is an example sketch from Matlab to just have Matlab read and send information to arduino by skipping the arduino IDE. I'm converting it to read the pressure sensor. I also want it to use the LoadCell data from the HX711 but since it's a 24-bit ADC we need the microcontroller to do some of the processing before sending the information to Matlab. Therefore I've included the HX711 add-on code to the Matlab folder. It doesn't seem to work, but this may be a problem with my Matlab install on the EB590 machine "EB475WS4.cecs.pdx.edu".