Updated MonoPamDataExplicit, moved figures
Updated MonoPamDataExplicit so that it was the same as MonoPamDataExplicit_compare, except the force calculation for 10mm only calls the simplified exponential force model (Fn4)  from normF10. Moved all of the pictures in the Figure Components folder into subfolders for easier organization. Should allow for easier type setting from editors, or deletion if needed. Also added a _deprecated subfolder for figures that I will probably delete but may still have a use for.