Extensor results improved
But I think the 42cm length needs kmax to be 345 or 340 mm, and the 48cm l_rest needs kmax to be down to 405 or back to 398. I changed it based off of the readings, but muscle length row in the excel files should be taken with a grain of salt, since they were measured in place, which is harder than when doing it when the ends are free.